DEV: move plugin-specific test to plugin tests
