+++ app/assets/javascripts/discourse/tests/acceptance/composer-test.js
@@ -1300,32 +1300,6 @@ acceptance("Composer - default category not set", function (needs) {
 });
 // END: Default Composer Category tests
 
-acceptance("Composer - current time", function (needs) {
-  needs.user();
-
-  test("composer insert current time shortcut", async function (assert) {
-    await visit("/t/internationalization-localization/280");
-
-    await click("#topic-footer-buttons .btn.create");
-    assert.dom(".d-editor-input").exists("the composer input is visible");
-    await fillIn(".d-editor-input", "and the time now is: ");
-
-    const date = moment().format("YYYY-MM-DD");
-
-    await triggerKeyEvent(".d-editor-input", "keydown", ".", {
-      ...metaModifier,
-      shiftKey: true,
-    });
-
-    const inputValue = query("#reply-control .d-editor-input").value.trim();
-
-    assert.ok(
-      inputValue.startsWith(`and the time now is: [date=${date}`),
-      "it adds the current date"
-    );
-  });
-});
-
 acceptance("composer buttons API", function (needs) {
   needs.user();
   needs.settings({
